Move all the Akka stuff into the archivist
