+++ archivist/src/main/scala/uk/ac/wellcome/platform/archive/archivist/Archivist.scala
@@ -11,10 +11,9 @@ import grizzled.slf4j.Logging
 import uk.ac.wellcome.json.JsonUtil._
 import uk.ac.wellcome.messaging.sns.{NotificationMessage, SNSConfig}
 import uk.ac.wellcome.platform.archive.archivist.flow._
+import uk.ac.wellcome.platform.archive.archivist.messaging.MessageStream
 import uk.ac.wellcome.platform.archive.archivist.models.BagUploaderConfig
 import uk.ac.wellcome.platform.archive.common.config.models.Parallelism
-import uk.ac.wellcome.platform.archive.common.flows.SupervisedMaterializer
-import uk.ac.wellcome.platform.archive.common.messaging.MessageStream
 import uk.ac.wellcome.typesafe.Runnable
 
 import scala.concurrent.Future
