+++ archivist/src/main/scala/uk/ac/wellcome/platform/archive/archivist/Main.scala
@@ -5,8 +5,10 @@ import com.typesafe.config.Config
 import uk.ac.wellcome.messaging.sns.NotificationMessage
 import uk.ac.wellcome.messaging.typesafe.SNSBuilder
 import uk.ac.wellcome.platform.archive.archivist.builders.TransferManagerBuilder
-import uk.ac.wellcome.platform.archive.archivist.config.BagUploaderConfigBuilder
-import uk.ac.wellcome.platform.archive.common.config.builders.MessagingBuilder
+import uk.ac.wellcome.platform.archive.archivist.config.{
+  BagUploaderConfigBuilder,
+  MessagingBuilder
+}
 import uk.ac.wellcome.storage.typesafe.S3Builder
 import uk.ac.wellcome.typesafe.WellcomeTypesafeApp
 import uk.ac.wellcome.typesafe.config.builders.AkkaBuilder
