+++ archivist/src/main/scala/uk/ac/wellcome/platform/archive/archivist/flow/ArchiveAndNotifyNextFlow.scala
@@ -1,4 +1,5 @@
 package uk.ac.wellcome.platform.archive.archivist.flow
+
 import akka.NotUsed
 import akka.stream.scaladsl.Flow
 import com.amazonaws.services.s3.AmazonS3
@@ -6,7 +7,6 @@ import com.amazonaws.services.sns.AmazonSNS
 import uk.ac.wellcome.messaging.sns.SNSConfig
 import uk.ac.wellcome.platform.archive.archivist.models.TypeAliases.BagDownload
 import uk.ac.wellcome.platform.archive.archivist.models.BagUploaderConfig
-import uk.ac.wellcome.platform.archive.common.flows.FoldEitherFlow
 import uk.ac.wellcome.platform.archive.common.models.BagRequest
 import uk.ac.wellcome.platform.archive.common.models.error.ArchiveError
