+++ archivist/src/main/scala/uk/ac/wellcome/platform/archive/archivist/flow/ArchiveZipFileFlow.scala
@@ -11,6 +11,7 @@ import grizzled.slf4j.Logging
 import uk.ac.wellcome.json.JsonUtil._
 import uk.ac.wellcome.messaging.sns.SNSConfig
 import uk.ac.wellcome.platform.archive.archivist.bag.ArchiveJobCreator
+import uk.ac.wellcome.platform.archive.archivist.messaging.SnsPublishFlow
 import uk.ac.wellcome.platform.archive.archivist.models.{
   BagUploaderConfig,
   FileDownloadComplete
@@ -20,7 +21,6 @@ import uk.ac.wellcome.platform.archive.archivist.models.TypeAliases.{
   BagDownload
 }
 import uk.ac.wellcome.platform.archive.archivist.models.errors.ArchiveJobError
-import uk.ac.wellcome.platform.archive.common.messaging.SnsPublishFlow
 import uk.ac.wellcome.platform.archive.common.models.error.ArchiveError
 import uk.ac.wellcome.platform.archive.common.models.{
   BagRequest,
