+++ archivist/src/main/scala/uk/ac/wellcome/platform/archive/archivist/flow/DigestItemJobFlow.scala
@@ -5,10 +5,6 @@ import akka.stream.scaladsl.Flow
 import com.amazonaws.services.s3.AmazonS3
 import grizzled.slf4j.Logging
 import uk.ac.wellcome.platform.archive.archivist.models.DigestItemJob
-import uk.ac.wellcome.platform.archive.common.flows.{
-  FoldEitherFlow,
-  OnErrorFlow
-}
 import uk.ac.wellcome.platform.archive.common.models.error.ArchiveError
 
 object DigestItemJobFlow extends Logging {
