+++ archivist/src/main/scala/uk/ac/wellcome/platform/archive/archivist/flow/DiscardLeftFlow.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.archive.common.flows
+package uk.ac.wellcome.platform.archive.archivist.flow
 
 import akka.NotUsed
 import akka.stream.scaladsl.Flow
