+++ archivist/src/main/scala/uk/ac/wellcome/platform/archive/archivist/flow/NotificationMessageFlow.scala
@@ -5,11 +5,11 @@ import akka.stream.scaladsl.{Flow, Source}
 import com.amazonaws.services.sns.AmazonSNS
 import uk.ac.wellcome.json.JsonUtil._
 import uk.ac.wellcome.messaging.sns.{NotificationMessage, SNSConfig}
-import uk.ac.wellcome.platform.archive.common.config.models.Parallelism
-import uk.ac.wellcome.platform.archive.common.messaging.{
+import uk.ac.wellcome.platform.archive.archivist.messaging.{
   NotificationParsingFlow,
   SnsPublishFlow
 }
+import uk.ac.wellcome.platform.archive.common.config.models.Parallelism
 import uk.ac.wellcome.platform.archive.common.models.IngestBagRequest
 import uk.ac.wellcome.platform.archive.common.progress.models.{
   ProgressEvent,
