+++ archivist/src/main/scala/uk/ac/wellcome/platform/archive/archivist/messaging/MessageStream.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.archive.common.messaging
+package uk.ac.wellcome.platform.archive.archivist.messaging
 
 import akka.NotUsed
 import akka.actor.ActorSystem
