+++ archivist/src/main/scala/uk/ac/wellcome/platform/archive/archivist/messaging/NotificationParsingFlow.scala
@@ -1,11 +1,11 @@
-package uk.ac.wellcome.platform.archive.common.messaging
+package uk.ac.wellcome.platform.archive.archivist.messaging
 
 import akka.NotUsed
 import akka.stream.scaladsl.Flow
 import io.circe.Decoder
 import uk.ac.wellcome.json.JsonUtil._
 import uk.ac.wellcome.messaging.sns.NotificationMessage
-import uk.ac.wellcome.platform.archive.common.flows.ProcessLogDiscardFlow
+import uk.ac.wellcome.platform.archive.archivist.flow.ProcessLogDiscardFlow
 
 /** Parses a NotificationMessage as an instance of type T, and only
   * emits the successfully parsed results.
