+++ archivist/src/test/scala/uk/ac/wellcome/platform/archive/archivist/messaging/MessageStreamTest.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.archive.common.messaging
+package uk.ac.wellcome.platform.archive.archivist.messaging
 
 import java.util.concurrent.ConcurrentLinkedQueue
 
@@ -14,7 +14,7 @@ import uk.ac.wellcome.fixtures.TestWith
 import uk.ac.wellcome.json.JsonUtil._
 import uk.ac.wellcome.messaging.fixtures.Messaging
 import uk.ac.wellcome.messaging.fixtures.SQS.{Queue, QueuePair}
-import uk.ac.wellcome.platform.archive.common.fixtures.ArchiveMessaging
+import uk.ac.wellcome.platform.archive.archivist.fixtures.ArchiveMessaging
 
 class MessageStreamTest
     extends FunSpec
